🐛 BugFix: Fix Unwanted Halts [main]
Changes Made:
	- Add exit code checks to main script
	- Add `fzf` as requirement
	- Change position of functions in `install-apps.sh``

I missed some places in the script where, the script would unintentially
give out errors or halt the execution. This commit fixes those issues by
checking the exit codes after the menu script exits, to see if the
installation should be continued, or if it should exit with or without
saving the `install.list` file. It also adds `fzf` as a requirement, so
that the menu can be used properly. One minor change is that position of
the functions in `install-apps.sh` has been changed to be before the
executtion of the main installation as they were not known to the script
before that point.

Files Changed:
	modified:   klips.sh
	modified:   scripts/install-apps.sh

Date: 17/07/2025, 12:27 A.M. IST

All Done.

Signed-off-by: Kunal Kumar <kunalkumarchourasiya2021@gmail.com>