♻️ Refactor: Add Checks And Restrictions [main]
Changes Made:
	- Add global requirement check for `lsb-release` to ensure distribtuion
	  detection works.
	- Rename variables in `run_fzf` function for consistency.
	- Add directory variable to paths for unambiguous references.
	- Add checks for packages list variables to avoid empty install commands.
	- Add new entries to package lists.
	- Add ignore for `.vscode` directory.

Had to refactor some code to ensure consistency and clarity in the script.
Therewere problems with distro detection and package installation commands
for which hecks were to ensure that detection program was installed and
the package list ariables were not empty before proceeding with installation.
Also renamed and added some variables for clarity and consistency in the
code. Some new entries were added to the package lists for better coverage.

Files Changed:
	modified:   .gitignore
	modified:   klips.sh
	modified:   scripts/arch/archibald.sh
	modified:   scripts/arch/browsers.list
	modified:   scripts/arch/cli-tui.list
	modified:   scripts/arch/commands.list
	modified:   scripts/arch/miscellaneous.list
	modified:   scripts/install-apps.sh

Date: 20/07/2025, 05:46P.M. IST

All Done.

Signed-off-by: Kunal Kumar <kunalkumarchourasiya2021@gmail.com>